[INPLACE-25] user와 adminuser 분리
