+++ backend/src/main/java/team7/inplace/admin/user/persistence/AdminUserRepository.java
@@ -1,9 +1,9 @@
-package team7.inplace.user.persistence;
+package team7.inplace.admin.user.persistence;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import team7.inplace.user.domain.AdminUser;
+import team7.inplace.admin.user.domain.AdminUser;
 
 @Repository
 public interface AdminUserRepository extends JpaRepository<AdminUser, Long> {
