+++ backend/src/main/java/team7/inplace/admin/user/presentation/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package team7.inplace.admin.dto;
+package team7.inplace.admin.user.presentation.dto;
 
 public record LoginRequest(
     String username,
