+++ backend/src/main/java/team7/inplace/security/application/CustomUserDetailsService.java
@@ -4,19 +4,20 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import team7.inplace.admin.user.application.AdminUserService;
+import team7.inplace.admin.user.application.dto.AdminUserInfo;
 import team7.inplace.security.application.dto.CustomUserDetails;
 import team7.inplace.user.application.UserService;
-import team7.inplace.user.application.dto.UserCommand;
 
 @RequiredArgsConstructor
 public class CustomUserDetailsService implements UserDetailsService {
 
-    private final UserService userService;
+    private final AdminUserService adminUserService;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        UserCommand.AdminUserInfo adminUserInfo = userService.findAdminUserByUsername(username)
-            .orElseThrow(()-> new UsernameNotFoundException("User is not found"));
+        AdminUserInfo adminUserInfo = adminUserService.findAdminUserByUsername(username)
+            .orElseThrow(() -> new UsernameNotFoundException("User is not found"));
 
         return CustomUserDetails.makeUser(adminUserInfo);
     }
