+++ backend/src/main/java/team7/inplace/user/persistence/UserRepository.java
@@ -10,6 +10,4 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
-
-    boolean existsByUsername(String username);
 }
